FRGRealTime

Documentation for FRGRealTime.

Index

FRGRealTime.Epi

FRGRealTime.VImintqs

FRGRealTime.deltafun

FRGRealTime.dkF1Allintqs

FRGRealTime.dkF2Allintqs

FRGRealTime.dkVImintqs

FRGRealTime.flowpm

FRGRealTime.flowpm_intcostheqs

FRGRealTime.flowpp

FRGRealTime.flowpp_intcostheqs

FRGRealTime.loopfunpm

FRGRealTime.loopfunpp

FRGRealTime.loopfunppfix

FRGRealTime.star1fun

FRGRealTime.star1funpm

FRGRealTime.star2fun

FRGRealTime.star2funpm

FRGRealTime.star3fun

FRGRealTime.star3funpm

FRGRealTime.star4fun

FRGRealTime.Epi — Method

Epi(k,m)

compute $\sqrt{(x^2+m)}$

FRGRealTime.VImintqs — Method

VImintqs(p0, ps, k, T, Npi,mfun::Function,lampifun::Function)

compute $\int_0^{k}dq_s qs^2\int_{-1}^{1}d\cos\theta \mathrm{Im}V(q_0,k)$. In our code, we perform integration over kprim, q0 & qs does not involved, so qs=k, q0=Epi(k, mfun(k)).

Arguments

mfun::Function: $m^2(k)$, input from zero momentum result

lampifun::Function: $\lambda_{4\pi}(k)$, input from zero momentum result.

FRGRealTime.deltafun — Function

deltafun(x,dϵ=0.02)

compute $\delta$ function with the approximation $\frac{\epsilon}{\pi(\epsilon^2+x^2)}$ and $\epsilon$ is set to be $0.02$ by default

FRGRealTime.dkF1Allintqs — Method

dkF1Allintqs(p0, ps, qsmax, k, m, T)

$F_1$ is an odd function about $p_0$

when p0>0,dkF1Allintqs=flowpp_intcostheqs

when p0<0,dkF1Allintqs=-flowpp_intcostheqs(-p0)

FRGRealTime.dkF2Allintqs — Method

flowpm(p0, ps, k, m, T)

$F_2$ is an odd function about $p_0$

when p0>0,dkF2Allintqs=flowpm_intcostheqs

when p0<0,dkF2Allintqs=-flowpm_intcostheqs(-p0)

FRGRealTime.dkVImintqs — Method

dkVImintqs(p0, ps, q0, qsmax, k, m, T, Npi, lam4pik)

compute $\int_0^{qsmax}dq_s qs^2\int_{-1}^{1}d\cos\theta \tilde{\partial_k}\mathrm{Im}V(q_0)$.

dkV4piImintqs only contains $V(q_0)$, for $-q_0$, we have $\int d\cos\theta V(q_0)=\int d\cos\theta V(-q_0)$, so we need an extra $2$ at somewhere.

Arguments

qsmax: we integrate $q_s$ from $0$ to $k$, qsmax will set to k when we do the integration $dk'$, it should be distinguished from $k'$

m: mass square, it will be $m(k')$ when we do the integration $dk'$.

lam4pik: $\lambda_{4\pi}$, it will be $\lambda_{4\pi}(k')$ when we do the integration $dk'$ .

FRGRealTime.flowpm — Method

flowpm(p0, ps, k, m, T)

compute $-\frac{1}{\pi}\tilde{\partial_k}\Im I_{2, k}(p)$

FRGRealTime.flowpm_intcostheqs — Method

flowpm_intcostheqs(p0, ps, qsmax, k, m, T)

compute

\[\begin{aligned} &\int_0^{qsmax}\!\!dq_s q_s^2\int_{-1}^{1}\!\!d\cos\theta \tilde{\partial_k}F_2\left(\sqrt{p_s^2+q_s^2+2p_sq_s\cos\theta}\right)\\ &=\int_0^{qsmax}\!\!dq_s q_s^2\int_{-1}^{1}\!\!d\cos\theta \tilde{\partial_k}F_2\left(\sqrt{p_s^2+q_s^2-2*p_s*q_s\cos\theta}\right) \end{aligned}\]

FRGRealTime.flowpp — Function

flowpp(p0, ps, k, m, T,δ=0.02)

compute $-\frac{1}{\pi}\tilde{\partial_k}\Im I_{1, k}(p)$

To be noticed that, flowpp doesn't constains $\tilde{\partial_k}\mathcal{F}_4$, we will consider it seperately.

At $p_0=2E_{\pi,k}$, flowpp has a $\delta$ function contribution, we use a rectangle function with width $\delta$ to approximate.

FRGRealTime.flowpp_intcostheqs — Method

flowpp_intcostheqs(p0, ps, qsmax, k, m, T)

compute

\[\begin{aligned} &\int_0^{qsmax}\!\!dq_s q_s^2\int_{-1}^{1}\!\!d\cos\theta \tilde{\partial_k}F_1\left(\sqrt{p_s^2+q_s^2+2*p_s*q_s\cos\theta}\right)\\ &=\int_0^{qsmax}\!\!dq_s q_s^2\int_{-1}^{1}\!\!d\cos\theta \tilde{\partial_k}F_1\left(\sqrt{p_s^2+q_s^2-2*p_s*q_s\cos\theta}\right) \end{aligned}\]

FRGRealTime.loopfunpm — Method

loopfunpm(p0, ps, k, m, T)

compute $-\frac{1}{\pi}\Im I_{2, k}(p)$

FRGRealTime.loopfunpp — Method

loopfunpp(p0, ps, k, m, T)

compute $-\frac{1}{\pi}\Im I_{2, k}(p)$, but without $\mathcal{F}_4$ contribution

FRGRealTime.loopfunppfix — Function

loopfunppfix(p0, ps, k, m, T)

compute $-\frac{1}{\pi}\Im I_{1, k}(p)$

FRGRealTime.star1fun — Method

star1fun(qp, qm, ps, m, T)

compute $-\frac{1}{\pi}\mathcal{F}_{1}\left(q_{+}, q_{-}, p, \bar{m}_{\pi, k}^{2}\right)$

FRGRealTime.star1funpm — Method

star1funpm(qp, qm, ps, m, T)

compute $-\frac{1}{\pi}\mathcal{F}_{1}^{\prime}\left(q_{+}, q_{-}, p, \bar{m}_{\pi, k}^{2}\right)$

Compared with $\mathcal{F}$, the $\mathcal{F'}$ has subtracted vacuum contributions. it will be used in $\mathrm{Im} I_2$

FRGRealTime.star2fun — Method

star2fun(qp, qm, ps, k, m, T)

compute $-\frac{1}{\pi}\mathcal{F}_{2}\left(q_{+}, q_{-}, k, p, \bar{m}_{\pi, k}^{2}\right)$

FRGRealTime.star2funpm — Method

star2funpm(qp, qm, ps, k, m, T)

compute $-\frac{1}{\pi}\mathcal{F}^{\prime}_{2}\left(q_{+}, q_{-}, k, p, \bar{m}_{\pi, k}^{2}\right)$

FRGRealTime.star3fun — Method

star3fun(qp, ps, k, m, T)

compute $-\frac{1}{\pi}\mathcal{F}_{3}\left(q_{+},k, p, \bar{m}_{\pi, k}^{2}\right)$

FRGRealTime.star3funpm — Method

star3funpm(qp, ps, k, m, T)

compute $-\frac{1}{\pi}\mathcal{F}^{\prime}_{3}\left(q_{+}, k, p, \bar{m}_{\pi, k}^{2}\right)$

FRGRealTime.star4fun — Function

star4fun(p0, p, k, m, T)

compute $-\frac{1}{\pi}\mathcal{F}_{4}\left(p_{0}, k, p, \bar{m}_{\pi, k}^{2}\right)$