FRGRealTime

Documentation for FRGRealTime.

Index

FRGRealTime.Coeffgamm2

FRGRealTime.Coeffgamm2Simple

FRGRealTime.Epi

FRGRealTime.F1All

FRGRealTime.F2All

FRGRealTime.PvdkF1Tildeps

FRGRealTime.VImSimple

FRGRealTime.VImintqs

FRGRealTime.VImintqsSimple

FRGRealTime.VImintqs_delta1

FRGRealTime.VReintqs

FRGRealTime.deltafun

FRGRealTime.dkF1All

FRGRealTime.dkF1Allintqs

FRGRealTime.dkF1TildeintqsAll

FRGRealTime.dkF2All

FRGRealTime.dkF2Allintqs

FRGRealTime.dkF2TildeintqsAll

FRGRealTime.dkVIm

FRGRealTime.dkVImintqs

FRGRealTime.dkVReintqs

FRGRealTime.flowpm

FRGRealTime.flowpm_intcostheqs

FRGRealTime.flowpp

FRGRealTime.flowpp_intcostheqs

FRGRealTime.loopfunpm

FRGRealTime.loopfunpp

FRGRealTime.loopfunppfix

FRGRealTime.propImSimple

FRGRealTime.propImintqs_delta1_dq0_delta1

FRGRealTime.star1fun

FRGRealTime.star1funpm

FRGRealTime.star2fun

FRGRealTime.star2funpm

FRGRealTime.star3fun

FRGRealTime.star3funpm

FRGRealTime.star4fun

FRGRealTime.Coeffgamm2 — Method

Coeffgamm2(k, T, mfun::Spline1D)

The coefficient before the two point function flow.

FRGRealTime.Coeffgamm2Simple — Method

Coeffgamm2Simple(k, T, m)

The coefficient before the two point function flow.

FRGRealTime.Epi — Method

Epi(k,m)

compute $\sqrt{(x^2+m)}$

FRGRealTime.F1All — Method

F1All(p0, ps, k, m, T)

$F_1$ is an odd function about $p_0$

when p0>0,F1All=loopfunpp

when p0<0,dkF1All=-loopfunpp(-p0)

F1All doesn't contains the type-2 delta function

FRGRealTime.F2All — Method

F2All(p0, ps, k, m, T)

$F_2$ is an odd function about $p_0$

when p0>0,F2All=loopfunpm

when p0<0,F2All=-loopfunpm(-p0)

FRGRealTime.PvdkF1Tildeps — Method

PvdkF1Tildeps(p0, ps, qsmax, k, m, T; kwargs...)

Piecewise hints:

We only integrate the p0' domain which the flowpp_intcostheqs is non zero.

FRGRealTime.VImSimple — Method

VImSimple(p0, ps, q0, k, m, T, Npi, lam4pik)

compute $\mathrm{Im}V_k(q_0)$, the k dependence of $\lambda_{4\pi}$ and $m$ are neglected.

VImSimple only contains $V(q_0)$, no $V(-q_0)$

VImSimple contains type-1 delta function

VImSimple doesn't contains type-2 delta function

FRGRealTime.VImintqs — Method

VImintqs(p0, ps, k, T, Npi,UVScale,mfun::Function,lamfun::Function)

compute $\int_0^{k}dq_s qs^2\int_{-1}^{1}d\cos\theta \mathrm{Im}V(q_0,k)$. In our code, we perform integration over kprim, q0 & qs does not involved, so qs=k, q0=Epi(k, mfun(k)).

VImintqs don't have any delta function contribution, we include the type-1 delta function in VImintqs_delta1 separately.

Arguments

mfun::Function: $m^2(k)$, input from zero momentum result

lampifun::Function: $\lambda_{4\pi}(k)$, input from zero momentum result.

FRGRealTime.VImintqsSimple — Method

VImintqsSimple(p0, ps, k, T, Npi, m, lamda,UVScale)

compute $\int_0^kq_s^2dqs\int_{-1}^{1}d\cos\theta\mathrm{Im}V_k(q_0)$, the k dependence of $\lambda_{4\pi}$ and $m$ are neglected.

VImintqsSimple only contains $V(q_0)$, no $V(-q_0)$

VImintqsSimple contains type-1 delta function

VImintqsSimple doesn't contains type-2 delta function

FRGRealTime.VImintqs_delta1 — Method

VImintqs_delta1(p0, ps, k, T, Npi,IRScale,UVScale, mfun::Spline1D, lamfun::Spline1D)

compute type-1 delta function contribution in $\mathrm{Im}V_k$, we do the triple integral analytically.

FRGRealTime.VReintqs — Method

VReintqs(p0, ps, k, T, Npi,IRScale,UVScale, mfun, lamfun)

compute $\int_0^{k}dq_s qs^2\int_{-1}^{1}d\cos\theta \mathrm{Re}V(q_0,k)$. In our code, we perform integration over kprim, q0 & qs does not involved, so qs=k, q0=Epi(k, mfun(k)).

VReintqs contains type-1 and type-2 delta function.

Arguments

mfun::Function: $m^2(k)$, input from zero momentum result

lampifun::Function: $\lambda_{4\pi}(k)$, input from zero momentum result.

FRGRealTime.deltafun — Function

deltafun(x,dϵ=0.02)

compute $\delta$ function with the approximation $\frac{\epsilon}{\pi(\epsilon^2+x^2)}$ and $\epsilon$ is set to be $0.02$ by default

FRGRealTime.dkF1All — Function

dkF1All(p0, ps, k, m, T)

$F_1$ is an odd function about $p_0$

when p0>0,dkF1All=flowpp

when p0<0,dkF1All=-flowpp(-p0)

dkF1All doesn't contains the type-2 delta function

FRGRealTime.dkF1Allintqs — Method

dkF1Allintqs(p0, ps, qsmax, k, m, T)

$F_1$ is an odd function about $p_0$

when p0>0,dkF1Allintqs=flowpp_intcostheqs

when p0<0,dkF1Allintqs=-flowpp_intcostheqs(-p0)

dkF1Allintqs doesn't contains the type-1 delta function

dkF1Allintqs doesn't contains the type-2 delta function

FRGRealTime.dkF1TildeintqsAll — Method

dkF1TildeintqsAll(p0, qsmax, k, m, T; kwargs...)

costh is not integrated we need an extra 2 at somewhere

FRGRealTime.dkF2All — Method

dkF2All(p0, ps, k, m, T)

$F_2$ is an odd function about $p_0$

when p0>0,dkF2All=flowpm

when p0<0,dkF2All=-flowpm(-p0)

FRGRealTime.dkF2Allintqs — Method

flowpm(p0, ps, k, m, T)

$F_2$ is an odd function about $p_0$

when p0>0,dkF2Allintqs=flowpm_intcostheqs

when p0<0,dkF2Allintqs=-flowpm_intcostheqs(-p0)

FRGRealTime.dkF2TildeintqsAll — Method

dkF2TildeintqsAll(p0, qsmax, k, m, T; kwargs...)

costh is not integrated we need an extra 2 at somewhere

FRGRealTime.dkVIm — Method

dkVIm(p0, ps, q0, k, m, T, Npi, lam4pik)

compute $\tilde{\partial_k}\mathrm{Im}V_k(q_0)$.

dkVIm only contains $V(q_0)$, no $V(-q_0)$

dkVIm contains type-1 delta function

dkVIm doesn't contains type-2 delta function

FRGRealTime.dkVImintqs — Method

dkVImintqs(p0, ps, q0, qsmax, k, m, T, Npi, lam4pik)

compute $\int_0^{qsmax}dq_s qs^2\int_{-1}^{1}d\cos\theta \tilde{\partial_k}\mathrm{Im}V(q_0)$.

dkVImintqs only contains $V(q_0)$, for $-q_0$, we have $\int d\cos\theta V(q_0)=\int d\cos\theta V(-q_0)$, so we need an extra $2$ at somewhere.

dkVImintqs doesn't have any delta function contribution, we include the type-1 delta function in VImintqs_delta1 separately.

Arguments

qsmax: we integrate $q_s$ from $0$ to $k$, qsmax will set to k when we do the integration $dk'$, it should be distinguished from $k'$

m: mass square, it will be $m(k')$ when we do the integration $dk'$.

lam4pik: $\lambda_{4\pi}$, it will be $\lambda_{4\pi}(k')$ when we do the integration $dk'$ .

FRGRealTime.dkVReintqs — Method

dkVReintqs(p0, ps, q0, qsmax, k, m, T, Npi, lam4pik)

compute $\int_0^{qsmax}dq_s qs^2\int_{-1}^{1}d\cos\theta \tilde{\partial_k}\mathrm{Re}V(q_0)$.

dkVImintqs only contains $V(q_0)$, for $-q_0$, we have $\int d\cos\theta V(q_0)=\int d\cos\theta V(-q_0)$, so we need an extra $2$ at somewhere.

dkVReintqs contains type-1 and type-2 delta function

Arguments

qsmax: we integrate $q_s$ from $0$ to $k$, qsmax will set to k when we do the integration $dk'$, it should be distinguished from $k'$

m: mass square, it will be $m(k')$ when we do the integration $dk'$.

lam4pik: $\lambda_{4\pi}$, it will be $\lambda_{4\pi}(k')$ when we do the integration $dk'$ .

FRGRealTime.flowpm — Method

flowpm(p0, ps, k, m, T)

compute $-\frac{1}{\pi}\tilde{\partial_k}\Im I_{2, k}(p)$

FRGRealTime.flowpm_intcostheqs — Method

flowpm_intcostheqs(p0, ps, qsmax, k, m, T)

compute

\[\begin{aligned} &\int_0^{qsmax}\!\!dq_s q_s^2\int_{-1}^{1}\!\!d\cos\theta \tilde{\partial_k}F_2\left(\sqrt{p_s^2+q_s^2+2p_sq_s\cos\theta}\right)\\ &=\int_0^{qsmax}\!\!dq_s q_s^2\int_{-1}^{1}\!\!d\cos\theta \tilde{\partial_k}F_2\left(\sqrt{p_s^2+q_s^2-2*p_s*q_s\cos\theta}\right) \end{aligned}\]

FRGRealTime.flowpp — Function

flowpp(p0, ps, k, m, T,δ=0.02)

compute $-\frac{1}{\pi}\tilde{\partial_k}\Im I_{1, k}(p)$

To be noticed that, flowpp doesn't contains $\tilde{\partial_k}\mathcal{F}_4$, we will consider it separately.

At $p_0=2E_{\pi,k}$, flowpp has a $\delta$ function contribution, we use a rectangle function with width $\delta$ to approximate.

FRGRealTime.flowpp_intcostheqs — Method

flowpp_intcostheqs(p0, ps, qsmax, k, m, T)

compute

\[\begin{aligned} &\int_0^{qsmax}\!\!dq_s q_s^2\int_{-1}^{1}\!\!d\cos\theta \tilde{\partial_k}F_1\left(\sqrt{p_s^2+q_s^2+2*p_s*q_s\cos\theta}\right)\\ &=\int_0^{qsmax}\!\!dq_s q_s^2\int_{-1}^{1}\!\!d\cos\theta \tilde{\partial_k}F_1\left(\sqrt{p_s^2+q_s^2-2*p_s*q_s\cos\theta}\right) \end{aligned}\]

FRGRealTime.loopfunpm — Method

loopfunpm(p0, ps, k, m, T)

compute $-\frac{1}{\pi}\Im I_{2, k}(p)$

FRGRealTime.loopfunpp — Method

loopfunpp(p0, ps, k, m, T)

compute $-\frac{1}{\pi}\Im I_{2, k}(p)$, but without $\mathcal{F}_4$ contribution

FRGRealTime.loopfunppfix — Function

loopfunppfix(p0, ps, k, m, T)

compute $-\frac{1}{\pi}\Im I_{1, k}(p)$

FRGRealTime.propImSimple — Method

propImSimple(p0, ps, T,IRScale,UVScale, Npi, m, lamda)

x[1] is qs, x[2] is costh, x[3] is k

Arguments

m: mass square, it's a constant number.

lamda: $\lambda_{4\pi}$, it's a constant number.

FRGRealTime.propImintqs_delta1_dq0_delta1 — Method

propImintqs_delta1_dq0_delta(p0, ps, T, IRScale, UVScale, Npi, mfun, lamfun)

Compute the delta function part appears

FRGRealTime.star1fun — Method

star1fun(qp, qm, ps, m, T)

compute $-\frac{1}{\pi}\mathcal{F}_{1}\left(q_{+}, q_{-}, p, \bar{m}_{\pi, k}^{2}\right)$

FRGRealTime.star1funpm — Method

star1funpm(qp, qm, ps, m, T)

compute $-\frac{1}{\pi}\mathcal{F}_{1}^{\prime}\left(q_{+}, q_{-}, p, \bar{m}_{\pi, k}^{2}\right)$

Compared with $\mathcal{F}$, the $\mathcal{F'}$ has subtracted vacuum contributions. it will be used in $\mathrm{Im} I_2$

FRGRealTime.star2fun — Method

star2fun(qp, qm, ps, k, m, T)

compute $-\frac{1}{\pi}\mathcal{F}_{2}\left(q_{+}, q_{-}, k, p, \bar{m}_{\pi, k}^{2}\right)$

FRGRealTime.star2funpm — Method

star2funpm(qp, qm, ps, k, m, T)

compute $-\frac{1}{\pi}\mathcal{F}^{\prime}_{2}\left(q_{+}, q_{-}, k, p, \bar{m}_{\pi, k}^{2}\right)$

FRGRealTime.star3fun — Method

star3fun(qp, ps, k, m, T)

compute $-\frac{1}{\pi}\mathcal{F}_{3}\left(q_{+},k, p, \bar{m}_{\pi, k}^{2}\right)$

FRGRealTime.star3funpm — Method

star3funpm(qp, ps, k, m, T)

compute $-\frac{1}{\pi}\mathcal{F}^{\prime}_{3}\left(q_{+}, k, p, \bar{m}_{\pi, k}^{2}\right)$

FRGRealTime.star4fun — Function

star4fun(p0, p, k, m, T)

compute $-\frac{1}{\pi}\mathcal{F}_{4}\left(p_{0}, k, p, \bar{m}_{\pi, k}^{2}\right)$